backend/requirements.txt
flask
flask-cors
python-dotenv
bleak
requests
google-auth
